+++ hata/discord/channel.py
@@ -14,12 +14,12 @@ from ..backend.event_loop import LOOP_TIME
 from ..backend.export import export, include
 
 from .bases import DiscordEntity, IconSlot, ICON_TYPE_NONE, maybe_snowflake
-from .core import CHANNELS
+from .core import CHANNELS, MESSAGES
 from .permission import Permission, PermissionOverwrite
 from .permission.permission import PERMISSION_NONE, PERMISSION_ALL, PERMISSION_PRIVATE, PERMISSION_PRIVATE_BOT, \
     PERMISSION_GROUP, PERMISSION_GROUP_OWNER, PERMISSION_TEXT_DENY, PERMISSION_VOICE_DENY, PERMISSION_STAGE_MODERATOR, \
     PERMISSION_VOICE_DENY_CONNECTION, PERMISSION_TEXT_AND_VOICE_DENY, PERMISSION_TEXT_AND_STAGE_DENY
-from .message import Message, MESSAGES
+from .message import Message
 from .user import User, ZEROUSER, create_partial_user_from_id, thread_user_create
 from .core import GC_CYCLER
 from .webhook import Webhook, WebhookRepr